requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
Kivy==2.3.1
Kivy-Garden==0.1.5
opencv-python==4.12.0.88
Pillow==11.3.0
docutils==0.22
pygments==2.19.2
numpy==2.2.6
filetype==1.2.0
openai-whisper
sounddevice
torch
